feat: commit gulp-app demo
