+++ demo/gulp-app/package.json
@@ -24,6 +24,16 @@
   "devDependencies": {
     "@icebreakers/eslint-config-ts": "^1.0.4",
     "@types/gulp": "^4.0.10",
+    "@types/gulp-debug": "^2.0.32",
+    "@types/gulp-if": "^3.0.0",
+    "@types/gulp-plumber": "^0.0.33",
+    "@types/gulp-postcss": "^8.0.2",
+    "@types/gulp-rename": "^2.0.1",
+    "@types/gulp-replace": "^1.1.0",
+    "@types/gulp-sass": "^5.0.0",
+    "@types/gulp-util": "^3.0.37",
+    "@types/node": "^18.15.13",
+    "@types/yargs": "^17.0.24",
     "autoprefixer": "^10.4.14",
     "del": "^6.1.1",
     "eslint": "^8.38.0",
@@ -32,22 +42,23 @@
     "gulp-debug": "^4.0.0",
     "gulp-if": "^3.0.0",
     "gulp-newer": "^1.3.0",
+    "gulp-plumber": "^1.2.1",
     "gulp-postcss": "^9.0.1",
     "gulp-qcloud-upload": "^2.2.0",
     "gulp-rename": "^2.0.0",
     "gulp-replace": "^1.0.0",
     "gulp-sass": "^5.1.0",
     "gulp-sftp": "^0.1.5",
+    "gulp-typescript": "^6.0.0-alpha.1",
     "gulp-util": "^3.0.8",
     "miniprogram-api-typings": "^3.9.1",
     "postcss": "^8.4.23",
-    "postcss-font-base64": "^1.0.4",
-    "postcss-load-config": "^4.0.1",
-    "postcss-px2rpx": "0.0.4",
+    "postcss-pxtransform": "^3.6.5",
     "postcss-rem-to-responsive-pixel": "^5.1.3",
     "prettier": "^2.8.7",
     "sass": "^1.62.0",
     "tailwindcss": "^3.3.1",
+    "ts-node": "^10.9.1",
     "typescript": "^5.0.4",
     "weapp-ide-cli": "^1.0.1",
     "weapp-tailwindcss-webpack-plugin": "^2.1.1",
