--- demo/gulp-app/src/app.js
@@ -1,8 +1,0 @@
-//app.js
-App({
-	onLaunch: function () {
-	},
-	globalData: {
-		userInfo: null
-	}
-})
