+++ demo/gulp-app/src/app.scss
@@ -1,83 +1,74 @@
-
 @tailwind base;
 
 @tailwind utilities;
 
 // @tailwind components;
-@import "assets/scss/support";
-@import "assets/scss/utils";
-// 雪碧图
-// @lazysprite "filetype";
-
-.test {
-	background-image: url(%ASSETS_IMG%/qr-r.jpg);
-}
 
 .app {
-	height: 100%;
-	justify-content: space-between;
-	text-align: center;
-	box-sizing: border-box;
-	font-family: -apple-system-font, Helvetica Neue, Helvetica, sans-serif;
+  height: 100%;
+  justify-content: space-between;
+  text-align: center;
+  box-sizing: border-box;
+  font-family: -apple-system-font, Helvetica Neue, Helvetica, sans-serif;
 }
 
 .app__inner {
-	margin: 20px;
+  margin: 20px;
 }
 
 .app__title {
-	font-size: $fontSize;
-	line-height: $fontSize + 4px;
-	font-weight: bold;
-	padding-bottom: 10px;
-	margin-bottom: 20px;
-	text-align: left;
-	border-bottom: .5px solid #EEEEEE;
+  font-size: 16px;
+  line-height: 22px;
+  font-weight: bold;
+  padding-bottom: 10px;
+  margin-bottom: 20px;
+  text-align: left;
+  border-bottom: 0.5px solid #eeeeee;
 }
 
 .app__desc {
-	font-size: $fontSizeSmaller;
-	// line-height: $fontSizeSmaller +2px;
-	text-align: left;
-	margin-bottom: 20px;
-	color: $colorGray;
+  font-size: 12px;
+  // line-height: $fontSizeSmaller +2px;
+  text-align: left;
+  margin-bottom: 20px;
+  @apply text-gray-400;
 }
 
 // web font icon
 @font-face {
-	font-family: 'fontello';
-	src: url("assets/fonts/fontello.ttf") format('truetype');
-	font-weight: normal;
-	font-style: normal;
+  font-family: "fontello";
+  src: url("assets/fonts/fontello.ttf") format("truetype");
+  font-weight: normal;
+  font-style: normal;
 }
 
-[class^="icon-"]:before, [class*=" icon-"]:before {
-	font-family: "fontello";
-	font-style: normal;
-	font-weight: normal;
-	speak: none;
-	display: inline-block;
-	text-decoration: inherit;
-	width: 1em;
-	margin-right: 0.2em;
-	text-align: center;
-	font-variant: normal;
-	text-transform: none;
-	line-height: 1em;
-	margin-left: 0.2em;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
+[class^="icon-"]:before,
+[class*=" icon-"]:before {
+  font-family: "fontello";
+  font-style: normal;
+  font-weight: normal;
+  speak: none;
+  display: inline-block;
+  text-decoration: inherit;
+  width: 1em;
+  margin-right: 0.2em;
+  text-align: center;
+  font-variant: normal;
+  text-transform: none;
+  line-height: 1em;
+  margin-left: 0.2em;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 }
 
 .icon-info:before {
-	content: '\e81b';
+  content: "\e81b";
 }
 
 .icon-date:before {
-	content: '\e829';
+  content: "\e829";
 }
 
 .icon-view:before {
-	content: '\e84f';
+  content: "\e84f";
 }
-
