+++ demo/gulp-app/src/app.ts
@@ -1,0 +1,7 @@
+// app.js
+App({
+  onLaunch: function () {},
+  globalData: {
+    userInfo: null
+  }
+})
