--- demo/gulp-app/src/assets/scss/_support.scss
@@ -1,15 +1,0 @@
-/**
- * ------------------------------------------------------------------
- * 支持文件
- * 需要引用的地方均需要加上这个支持文件
- *
- * ------------------------------------------------------------------
- *
- * @author  JeffMa
- * @link    https://devework.com/
- * @data    2017-06-11
- */
-
-@import "mixins";
-
-@import "var";
