--- demo/gulp-app/src/assets/scss/_var.scss
@@ -1,44 +1,0 @@
-/**
- * ------------------------------------------------------------------
- * Sass 变量
- *
- * ------------------------------------------------------------------
- *
- * @author  JeffMa
- * @link    https://devework.com/
- * @data    2017-06-11
- */
-
-// Margin
-$marginTopSmaller: 10px;
-$marginTopSmall: 15px;
-$marginTopMedium: 20px;
-$marginTopLarge: 30px;
-$marginTopLarger: 40px;
-
-// Padding
-$paddingTopSmaller: 10px;
-$paddingTopSmall: 15px;
-$paddingTopMedium: 20px;
-$paddingTopLarge: 30px;
-$paddingTopLarger: 40px;
-
-// Color
-$colorBlue: #20A0FF;
-$colorGreen: #13CE66;
-$colorGray: #475669;
-$colorBlack: #000;
-$colorRed: #FF4949;
-$colorYellow: #F7BA2A;
-
-$color: #787878;
-$colorLink: #1D8CE0;
-
-$backGroundColor: #fff;
-
-// Font
-$fontSize: 16px;
-$fontSizeSmall: 14px;
-$fontSizeSmaller: 12px;
-$fontSizeLarge: 18px;
-$fontSizeLarger: 22px;
