+++ demo/gulp-app/src/pages/index/index.scss
@@ -1,5 +1,5 @@
 // 支持文件
-@import "../../assets/scss/support";
+
