+++ demo/gulp-app/src/pages/index/index.wxml
@@ -2,27 +2,25 @@
 <view class="h-full text-center">
   <view class="m-[20px]">
     <view>
-      <image src="https://res.jianhui.org/weapp-workflow/weapp-worfkflow@2x.jpg" mode="widthFix" class="w-full h-auto"></image>
+      <image src="../../image/logo.jpg" mode="widthFix" class="w-full h-auto"></image>
     </view>
-    <view>
-      <view class="text-[26px] mt-[20px] font-medium">WeApp-Workflow</view>
-      <view class="text-[14px] mt-[20px]">基于Gulp 的微信小程序前端开发工作流</view>
-      <view class="text-[14px] mt-[20px] text-[#5ba4e5]" bindtap="copyText" data-text="https:\/\/github.com/Jeff2Ma/WeApp-Workflow">
-        https://github.com/Jeff2Ma/WeApp-Workflow
+    <view class="space-y-4">
+      <view class="text-[14px]">基于Gulp 的微信小程序前端开发工作流</view>
+      <view class="text-[14px] text-[#5ba4e5]" bindtap="copyText" data-text="https://github.com/sonofmagic/weapp-tailwindcss-webpack-plugin">
+        https://github.com/sonofmagic/weapp-tailwindcss-webpack-plugin
       </view>
+      <view class="text-xs font-medium">Fork from Jeff2Ma/WeApp-Workflow</view>
     </view>
     <view class="{{contentStyle}}">
-      <view class="text-base font-bold pb-[10px] mb-[20px] text-left border-b border-[#EEEEEE]">功能介绍</view>
-      <text class="pl-[15px] text-left block">
-        <text class="u-d-b u-li">SCSS 实时编译为 WXSS</text>
-        <text class="u-d-b u-li">WXSS(CSS) 中px 单位转小程序单位rpx</text>
-        <text class="u-d-b u-li">图片压缩</text>
-        <text class="u-d-b u-li">自动上传图片到CDN 并更新路径为https 绝对路径</text>
-        <text class="u-d-b u-li">Font 文件转为base64 编码</text>
-        <text class="u-d-b u-li">全自动构建雪碧图及生成相应CSS</text>
-      </text>
-      
-      <!-- <image src="%ASSETS_IMG%/qr-r@2x.png"></image> -->
+      <view class="text-base font-bold pb-[10px] mb-[20px] text-left border-b border-[#EEEEEE]">
+        功能介绍
+      </view>
+      <view class="pl-[15px] text-left block">
+        <view class="text-xs">支持 tailwindcss</view>
+        <view class="text-xs">SCSS 实时编译为 WXSS</view>
+        <view class="text-xs">px 和 rem 单位转小程序单位rpx</view>
+        <view class="text-xs">支持 typescript</view>
+      </view>
     </view>
   </view>
 </view>
