--- demo/gulp-app/src/pages/more/more.js
@@ -1,8 +1,0 @@
-// logs.js
-const util = require('../../utils/util.js')
-Page({
-  data: {
-    logs: []
-  },
-  onLoad: function () {}
-})
