+++ demo/gulp-app/src/pages/more/more.ts
@@ -1,0 +1,11 @@
+// logs.js
+import { formatTime } from '../../utils/util'
+
+Page({
+  data: {
+    logs: []
+  },
+  onLoad: function () {
+    console.log(formatTime(new Date()))
+  }
+})
