+++ demo/gulp-app/src/utils/util.ts
@@ -1,0 +1,16 @@
+export function formatTime(date: Date) {
+  const year = date.getFullYear()
+  const month = date.getMonth() + 1
+  const day = date.getDate()
+
+  const hour = date.getHours()
+  const minute = date.getMinutes()
+  const second = date.getSeconds()
+
+  return [year, month, day].map(formatNumber).join('/') + ' ' + [hour, minute, second].map(formatNumber).join(':')
+}
+
+export function formatNumber(n: number | string) {
+  n = n.toString()
+  return n[1] ? n : '0' + n
+}
